refactor : service repository packaging , OrderStatus impl 추가 인터페이스 추가
