+++ src/main/java/com/sparta/nalda/dto/order/OwnerOrderResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.nalda.dto;
+package com.sparta.nalda.dto.order;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
