+++ src/main/java/com/sparta/nalda/service/order/OrderService.java
@@ -1,0 +1,23 @@
+package com.sparta.nalda.service.order;
+
+import com.sparta.nalda.dto.order.OrderListResponseDto;
+import com.sparta.nalda.dto.order.OrderRequestDto;
+import com.sparta.nalda.dto.order.OrderResponseDto;
+import com.sparta.nalda.dto.order.OwnerOrderResponseDto;
+import com.sparta.nalda.entity.OrderEntity;
+
+import java.util.List;
+
+
+public interface OrderService {
+
+  OrderEntity createOrder(OrderRequestDto requestDto);
+
+  OrderResponseDto findById(Long id);
+
+  List<OrderListResponseDto> findAllOrderList();
+
+  List<OwnerOrderResponseDto> findAllOrderListOwner(Long ownerId);
+
+
+}
