+++ src/main/java/com/sparta/nalda/util/OrderStatus.java
@@ -2,8 +2,8 @@ package com.sparta.nalda.util;
 
 public enum OrderStatus {
     PENDING, // 보류
-    ACCEPTED, //수락
     IN_PROGRESS, //조리 중
     DELIVERING, //배달 중
-    CANCELLED //취소
+    CANCELLED, //취소
+    DELIVERED // 배달 완료
 }
